==> Preprocessing/5.c.Residuals_datafilteringcohort_hc.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns 
import matplotlib.pyplot as plt
import statsmodels.formula.api as smf

residuals = pd.read_csv("Residualised_healthycontrol_training_all.csv")


# Load residuals
residuals = pd.read_csv("Residualised_healthycontrol_training_all.csv")
residualstesting = pd.read_csv("Testing_data_residualstraininghealthycontrol.csv")
# Training sets
training_all = pd.read_csv("Training_all.csv")  # has ["eid", "Diagnosis"]
training_eids = training_all["eid"].tolist()
diagnosis_dict = dict(zip(training_all["eid"], training_all["Diagnosis"]))

training_neuro_eids = pd.read_csv("Training_neurodegenerative.csv")["eid"].tolist()
training_hc_eids = pd.read_csv("Training_healthycontrol.csv")["eid"].tolist()
training_prodromals_eids = pd.read_csv("Training_all_prodromals.csv")["eid"].tolist()
training_baselinepd_eids = pd.read_csv("Training_baselinePD.csv")["eid"].tolist()


# Testing set (has its own diagnosis column)
testing_all = pd.read_csv("Testing_all.csv")  
testing_eids = testing_all["eid"].tolist()
testing_dict = dict(zip(testing_all["eid"], testing_all["Diagnosis"]))

# Subset residuals


residuals_neuro_training = residuals[residuals["eid"].isin(training_neuro_eids)]
residuals_hc_training = residuals[residuals["eid"].isin(training_hc_eids)]
residuals_training_all = residuals[residuals["eid"].isin(training_eids)]
residuals_testing = residualstesting[residualstesting["eid"].isin(testing_eids)]
residuals_prodromals_training = residuals[residuals["eid"].isin(training_prodromals_eids)]
residuals_baselinepd = residuals[residuals["eid"].isin(training_baselinepd_eids)]

# Add diagnosis to training splits
for df in [residuals_neuro_training, residuals_hc_training, residuals_training_all,residuals_prodromals_training,residuals_baselinepd]:
    df["Diagnosis"] = df["eid"].map(diagnosis_dict)

# Add diagnosis to testing split (different source)
residuals_testing["Diagnosis"] = residuals_testing["eid"].map(testing_dict)

residuals_neuro_training.to_csv("Training_residualshealthycontrol_neurodegenerative.csv", index=False)
residuals_hc_training.to_csv("Training_residualshealthycontrol_healthycontrol.csv", index=False)
residuals_training_all.to_csv("Training_residualshealthycontrol_all.csv", index=False)
residuals_testing.to_csv("Testing_residualshealthycontrol_all.csv", index=False)
residuals_prodromals_training.to_csv("Training_prodromals_residualshealthycontrol.csv", index=False)
residuals_baselinepd.to_csv("Training_residualshealthycontrol_baselinepd.csv", index=False)

residuals_baselinepd
